requirements.txt
altair==4.2.0
branca==0.5.0
folium==0.12.1.post1
geopy==2.2.0
numpy==1.21.5
pandarallel==1.6.1
pandas==1.4.2
Pillow==9.2.0
plotly==5.6.0
protobuf
seaborn==0.11.2
streamlit==1.11.0
streamlit_folium==0.6.13
google-cloud-storage==2.4.0
swifter
modin 
